Add image segmentation
Tesseract makebox task doesn't arrange characters to words and lines, so an algorithm was added to deal with this task.
This is it's initial version, which seem to perform quite good. Space width is currently specified as a constant and the function contains some debugging, which will be removed as soon as the algorithm is fully tested and integrated.

There was also a fix of the ImageProcessorConfiguration class, which for some misterious reasons contained weird values in its default initialization.