Add graphical interface and speech
A graphical interface was added using wxpython. Currently, it allows the user to load and recognize an image, place each border to the cursor position and also remove it. The About menu entry is already present, but not implemented yet.
LinuxSpeech class was added to wrap speechd in a cross-platform way, the interface should allow implementation of other systems on other platforms.
speech section was added to the configuration to let the user tweak his / her parameters, SpeechConfiguration class represents it in code.
appsdir library was added to allow loading configuration from platform native configuration directories and is already in use.