Add borders switching, complete removing, output image processing and result formats configuration
Function switch horizontal and vertical borders was added to make possible selections from below or top of the edge lines to edges of the image.
Output image processing allows the user to process the cropped image before being send to Mathpix, although this option is turned off by default.
Speaking of what, there is now also another option in the image processing configuration called active, which directly turns the processing on or off.

An option to remove all borders at once was added as this action seems to be performed quite a lot.

And also, user can now specify the resulting Mathpix format, currently supported are Asciimath and LaTeX (latex simplified).

As a minor change, the program now tries various locations for loading configuration, including the current directory. This change should make math_scanner more portable, although there is still not an option for setting up path to Tesseract.

And also in the tune of Tesseract, the configuration has been actually implemented, so setting things up in settings.yaml should work now.