Add columns support
Added an ability to split the image into columns based on a vertical middle line crossing a selected region.
Any number of columns can be created. math_scanner's _columns currently holds them as tuples of the column image, boxes and text, this would perhaps be better suited for a class, but I can't think out any logical name for it.
The new capabilities include:
* Split to columns
* Switch to previous / next column
* Cancel columns (return to the original image)