Fix borders switching, remove recognize full image menu entry
Borders switching was mistakenly implemented in a wrong ways, switching vertical borders in horizontal option and vice versa. This is fixed now.
The Recognize full image option was removed, as it was practically useless due to not set borders aligning to image edges.
Start of the program is now also conditioned by the if __name__=="__main__" condition. This allows the code to be used as a library, what could be useful for testing and debugging purposes.